ai-assisted-report/requirements.txt
openai>=1.30.0
python-dotenv
pandas
matplotlib
reportlab
